WIP: Reenvision Play page, splitting apart quick and custom games
