+++ src/components/TogglePreferRequired/TogglePreferRequired.styl
@@ -1,0 +1,122 @@
+/*
+ * Copyright (C)  Online-Go.com
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+ 
+ toggle-prefer-required-height = 1.3rem
+ toggle-prefer-required-half-height = 0.65rem
+ toggle-prefer-required-width = 2.5rem
+ //toggle-prefer-required-half-width = 1.25rem
+
+.TogglePreferRequired-outer {
+    display: inline-flex;
+    flex-direction: row;
+    align-items: center;
+    //padding-top: 1.1rem;
+    user-select: none;
+
+    
+    .priority {
+        font-size: small;
+    }
+    
+    .handle {
+        cursor: pointer;
+    }
+}
+
+.TogglePreferRequired-container {
+    display: inline-flex;
+    flex-direction: column;
+    align-items: center;
+    //padding-top: 1.1rem;
+    user-select: none;
+
+    
+    .priority {
+        font-size: small;
+    }
+}
+
+.TogglePreferRequired {
+    flex-grow: 0;
+    flex-shrink: 0;
+    display: inline-block;
+    height: toggle-prefer-required-height;
+    width: "calc(%s * 2)" % toggle-prefer-required-width;
+    border-radius: toggle-prefer-required-half-height;
+    position: relative;
+    themed background-color shade1
+    white-space: nowrap;
+    
+    .mid {
+        display: inline-block;
+        width: toggle-prefer-required-width;
+        height: toggle-prefer-required-height;
+        margin: 0 -0.02rem;
+        border-radius: toggle-prefer-required-half-height;
+        animation: 0.3s
+        transition: background-color 0.5s;
+
+        &.active {
+            themed background-color primary
+        }
+        
+        &.left.prefer {
+            border-top-right-radius: 0;
+            border-bottom-right-radius: 0;
+        }
+        &.right.prefer {
+            border-top-left-radius: 0;
+            border-bottom-left-radius: 0;
+        }
+    }
+    .handle {
+        display: inline-block;
+        border-radius: 50%;
+        border: 0;
+        margin: 0;
+        background-color: white;
+        position: absolute;
+        width: "calc(%s - 2px)" % toggle-prefer-required-height;
+        height: "calc(%s - 2px)" % toggle-prefer-required-height;
+        top: 1px;
+        transition: transform 0.3s;
+        left: 50%;
+
+        &.left {
+            transform: translateX(-200%);
+        }
+        &.right {
+            transform: translateX(100%);
+        }
+        
+        &.indifferent {
+            transform: translateX(-50%);
+            themed background-color shade0
+            width: "calc(%s - 4px)" % toggle-prefer-required-height;
+            height: "calc(%s - 4px)" % toggle-prefer-required-height;
+            border: 1px solid white;
+        }
+
+        &.priority-indifferent {
+            background-color: white !important
+        }
+        
+    }
+    
+    cursor: pointer;
+}
