+++ src/views/Play/Play.styl
@@ -20,7 +20,11 @@
     max-width: 1110px;
 
     .play-column {
-        margin: 0 auto;
+        margin-top: 2rem;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        //justify-content: center;
     }
 
     .seek-graph-container {
@@ -40,34 +44,6 @@
         box-sizing: border-box;
     }
 
-    .automatch-container {
-        flex: 1;
-        display: flex;
-        flex-direction: column;
-        align-items: stretch;
-
-        .small-spinner {
-            width: 20px;
-            height: 20px;
-            position: relative;
-        }
-    }
-
-    .automatch-header {
-        flex-shrink: 0;
-        flex-grow: 0;
-        flex-basis: 2.5rem;
-        font-size: 1.3rem;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-
-        .btn-group {
-            display: flex;
-            flex-wrap: nowrap;
-        }
-    }
-
     .rengo-management-row {
         td {
             padding-bottom: 2rem;
@@ -80,7 +56,6 @@
         .size {
 
         }
-
     }
 
     // long specifier to make sure this specific column is not overriden
@@ -142,125 +117,6 @@
         justify-content: flex-end;
     }
 
-    .automatch-row-container {
-        flex-grow: 1;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .automatch-row {
-        display: flex;
-        justify-content: space-around;
-        align-content: stretch
-        width: 100%;
-        flex-wrap: wrap;
-
-        button {
-            font-size: 1.1rem;
-            font-weight: bold;
-            flex-grow: 1;
-            flex-basis: 13rem;
-            max-width: 13rem;
-            height: 4rem;
-            margin: 0.5rem;
-            white-space: nowrap;
-            display: flex;
-            align-items: center;
-            align-content: center;
-            justify-content: center;
-
-            .play-button-text-root {
-                position: relative;
-                height: 1.5rem;
-
-
-                .time-per-move {
-                    position: absolute;
-                    bottom: -1.2rem;
-                    left: 0;
-                    right: 0;
-                    text-align: center;
-                }
-            }
-
-            .fa, .ogs-turtle {
-                margin: 0.2rem;
-            }
-
-            .ogs-turtle {
-                position: relative;
-                width: 1.0rem;
-            }
-            .ogs-turtle::before {
-                position: absolute;
-                top: -0.5rem;
-                left: 0;
-            }
-
-            .time-per-move {
-                font-size: font-size-smaller;
-            }
-        }
-    }
-
-    .automatch-settings {
-        font-size: 1rem;
-        flex: 0;
-        text-align: left;
-        white-space: nowrap;
-
-        .automatch-settings-link {
-            cursor: pointer;
-            user-select: none;
-        }
-
-        i {
-            padding-right: 0.25rem;
-        }
-    }
-
-    .automatch-settings-corr {
-        flex: 0;
-        text-align: justify;
-        font-size: 1rem;
-        padding: .5rem;
-        padding-left: 3rem;
-    }
-
-    .custom-game-row {
-        display: flex;
-        justify-content: space-around;
-        align-content: stretch
-        width: 100%;
-
-        button {
-            font-size: 1.1rem;
-            font-weight: bold;
-            flex-grow: 1;
-            width: 100%;
-            max-width: 13rem;
-            margin: 0.5rem;
-            height: 4rem;
-
-            .fa {
-                margin: 0.2rem;
-            }
-        }
-    }
-
-    .custom-game-header {
-        display: flex;
-        justify-content: space-around;
-        align-content: stretch
-        width: 100%;
-        justify-content: space-between;
-        align-items: left;
-        font-size: 1.3rem;
-        margin-top: 1rem;
-    }
-
     .showall-selector {
         width: 100%;
         text-align: left;
@@ -270,6 +126,44 @@
 
     #rengo-table {
         width: 100%;
+    }
+    
+    .MiniGoban {
+        .small.board {
+            cursor: default !important;
+        }
+    }
+    
+    button.active {
+        themed-important border-color primary
+    }
+    
+    .tab-head {
+        display: flex;
+        flex-wrap: wrap;
+        align-items: center;
+        justify-content: space-between;
+    }
+
+    .tabs-container {
+        text-align: center;
+    }
+    .tab {
+        cursor: pointer;
+        margin-left: 0.5rem;
+        margin-right: 0.5rem;
+        padding: 0.4rem;
+        border-bottom: 2px solid transparent;
+        white-space: nowrap;
+        display: inline-block;
+
+        &:hover {
+            themed border-color shade2
+        }
+        &.active {
+            themed border-color shade1
+        }
+
     }
 }
 
@@ -287,6 +181,21 @@
     display: inline-block;
 }
 
+.custom-games-header {
+    display: flex;
+    column-span: all;
+    font-weight: bold;
+    white-space: nowrap;
+    padding-bottom: 0.5 rem;
+    font-size: 1.3rem;
+    align-items: center;
+    justify-content: space-between;
+    
+    button {
+
+    }
+}
+
 #challenge-list {
     display: table;
     width: 100%;
@@ -298,12 +207,6 @@
     }
 
     // Flex to escape from the containing table formats
-    .custom-games-list-header-row {
-        display: flex;
-        font-weight: bold;
-        white-space: nowrap;
-        padding-bottom: 0.5 rem;
-    }
 
     .challenge-row {
         padding-top: 0.2em;
